Restructure users feature
