--- frontend/src/features/users/create/index.ts
@@ -1,4 +1,0 @@
-export * from "./UserCreateDetailsPage";
-export * from "./UserCreateLayout";
-export * from "./UserCreateRolePage";
-export * from "./UserCreateTypePage";
