+++ frontend/src/features/users/hooks/useUserListRequest.ts
@@ -1,4 +1,4 @@
-import { useInitialApiGetWithErrors, USER_LIST_REQUEST_PATH, UserListResponse } from "@kiesraad/api";
+import { useInitialApiGetWithErrors, USER_LIST_REQUEST_PATH, UserListResponse } from "@/api";
 
 export function useUserListRequest() {
   const path: USER_LIST_REQUEST_PATH = `/api/user`;
