--- frontend/src/features/users/index.ts
@@ -1,3 +1,0 @@
-export * from "./UserListPage";
-export * from "./create";
-export * from "./update";
