--- frontend/src/features/users/update/index.ts
@@ -1,1 +1,0 @@
-export * from "./UserUpdatePage";
