+++ frontend/src/lib/util/hook/index.ts
@@ -3,4 +3,3 @@ export * from "./useForm";
 export * from "./useNumericParam";
 export * from "./usePollingStationStatus";
 export * from "./usePreventFormEnterSubmit";
-export * from "./useQueryParam";
